# openAPI/OilAndGasContract.yaml
openapi: 3.0.0
info:
  version: '3.0'
  title: Oil and Gas Reporting System (OGRS-API)
  description: >-
    This is an api  to allows users to obtain infomation about the oil and gas wells database and generating a punch of related reposts such as wells count, insurance as well as production allocation funcationality
  termsOfService: http://ourcompany.com/terms-of-use
  contact:
    name: Muhammed Elbasiony Wafa
    email: mohamed.wafa770@gmail.com
  license:
    name: reporting API liecence
    url: http://liecence.anything.com
servers:
  - url: http://www.ourcompany.com/v1
    description: Dev server

  - url: http://www.ourcompany.com/v2
    description: Production server
...
paths:
...
  /wells/{wellId}/labs:
    post:
      tags:
        - Lab
      summary: Add a new lab Measurement
      description: Add a new lab Measurement to well
      operationId: addLabMeasurement
      parameters:
        - name: wellId
          in: path
          description: ID of lab to return
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: LabMeasurement object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/labMeasurementRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/labMeasurementRequest'
        required: true
      responses:
        "204":
          description: success
        "400":
          description: un expected body
        "401":
          description: not authintecated user
        "500":
          description: un expected error
        "403":
          description: not authorized user
        "405":
          description: Invalid input
          content: {}

    get:
      tags:
        - Lab
      summary: Find labs between two dates
      description: Returns labs between two dates
      operationId: getAllLabsInWell
      parameters:
        - name: wellId
          in: path
          description: ID of lab to return
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: beginDate
          in: query
          description: getting the daily actions that are of that date or after
          required: false
          style: form
          explode: true
          schema:
            type: string
            example: 2021-03-16T00:00:00.000+0000
        - name: endDate
          in: query
          description: getting the daily actions that are of that date or after
          required: false
          style: form
          explode: true
          schema:
            type: string
            example: 2021-03-16T00:00:00.000+0000
      responses:
        "204":
          description: Successful pull of the lab measurements for the well with the given id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllLabsOfAWell'
        "400":
          description: Invalid ID supplied
          content: {}
        "404":
          description: Lab not found
          content: {}
        "401":
          description: not authintecated user
        "500":
          description: un expected error
        "403":
          description: not authorized user
  /wells/labs:
    get:
      tags:
        - Lab
      summary: Find labs between two dates
      description: Returns labs between two dates
      operationId: getAllLabs
      parameters:
        - name: beginDate
          in: query
          description: getting the daily actions that are of that date or after
          required: false
          style: form
          explode: true
          schema:
            type: string
            example: 2021-03-16T00:00:00.000+0000
        - name: endDate
          in: query
          description: getting the daily actions that are of that date or after
          required: false
          style: form
          explode: true
          schema:
            type: string
            example: 2021-03-16T00:00:00.000+0000
      responses:
        "204":
          description: Successful pull of the lab measurements for the well with the given id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllLabs'
        "400":
          description: Invalid ID supplied
          content: {}
        "404":
          description: Lab not found
          content: {}
        "401":
          description: not authintecated user
        "500":
          description: un expected error
        "403":
          description: not authorized user
  /wells/{wellId}/labs/{labId}:
    put:
      tags:
        - Lab
      summary: Update a lab Measurement
      description: Update an existing lab Measurement in a well
      operationId: updateLabMeasurement
      parameters:
        - name: wellId
          in: path
          description: ID of lab to return
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: labId
          in: path
          description: record Id of the lab measurements table
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: LabMeasurement object that needs to be updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/labMeasurementRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/labMeasurementRequest'
        required: true
      responses:
        "204":
          description: success
        "400":
          description: un expected body
        "401":
          description: not authintecated user
        "500":
          description: un expected error
        "403":
          description: not authorized user
        "405":
          description: Invalid input
          content: {}

    delete:
      tags:
        - Lab
      summary: Deletes a lab
      description: Deletes a lab by its id
      operationId: deleteLabById
      parameters:
        - name: wellId
          in: path
          description: Delete a lab in a well by its id
          required: true
          style: simple
          explode: false
          schema:
            type: integer
        - name: labId
          in: path
          description: Delete a lab by its id
          required: true
          style: simple
          explode: false
          schema:
            type: integer
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Lab not found
        "401":
          description: not authintecated user
        "500":
          description: un expected error
        "403":
          description: not authorized user
          content: {}

    get:
      tags:
        - Lab
      summary: Find a certain lab on a certain well
      description: Returns a certain lab on a certain well
      operationId: getLabByWellIdAndLabId
      parameters:
        - name: wellId
          in: path
          description: ID of lab to return
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: labId
          in: path
          description: record Id of the lab measurements table
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/labMeasurementResponse'
        "400":
          description: Invalid ID supplied
          content: {}
        "404":
          description: Lab not found
          content: {}
        "401":
          description: not authintecated user
        "500":
          description: un expected error
        "403":
          description: not authorized user
...
  /wells/tests:
    get:
      tags:
        - wellTests
      summary: Return a list of all available well test records for all wells.
      description: returning a list of available test records
      operationId: findAllTests
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllTests'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server error.
  /wells/{id}/tests:
    get:
      tags:
        - wellTests
      summary: Find a list of  well Test records for specific well Id.
      description: Returing a list of all avilable well test records based on a given well id, For valid response try integer IDs with value >= 1 and <= 1000
      operationId: getTestById
      parameters:
        - name: id
          in: path
          description: ID of the well of which we need to retreive the tests records
          required: true
          schema:
            minimum: 1.0
            type: integer
        - name: beginDate
          in: query
          description: getting the test  measurements records that are of that date or after
          schema:
            type: string
            format: date-time
            example: 2021-03-16
        - name: endDate
          in: query
          description: getting the test  measurements records that are of that date or before
          schema:
            type: string
            format: date-time
            example: 2021-03-16
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllTests'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server error.

    post:
      tags:
        - wellTests
      summary: Add a new test record to the database for specific well
      operationId: addTestRecord
      parameters:
        - name: id
          in: path
          description: ID of the well of which we need to add the tests records
          required: true
          schema:
            minimum: 1.0
            type: integer
      requestBody:
        description: test roew object that needs to be added to the database
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/wellTestRequest'
        required: true
      responses:
        200:
          description: Record created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/wellTestResponse'
        400:
          description: Bad request. Must contain test record object in json format.
        401:
          description: Authorization information is missing or invalid.

        405:
          description: Invalid input. must be in json format
          content: {}
        500:
          description: Internal server Error
  /wells/{id}/tests/{recordId}:
    put:
      tags:
        - wellTests
      summary: Update an existing test record based on a given Id
      operationId: updateWellTestRecord
      parameters:
        - name: id
          in: path
          description: ID of the well of which we need to update the record.
          required: true
          schema:
            minimum: 1.0
            type: integer
        - name: recordId
          in: path
          description: ID of the well test record that needs to be deleted
          required: true
          schema:
            minimum: 1.0
            type: integer
      requestBody:
        description: well test row object that needs to be added to the database
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/wellTestRequest'

      responses:
        200: #can return 201 with empty body
          description: Ok, returns the updated test record object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/wellTestResponse'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Authorization information is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server error.

    delete:
      tags:
        - wellTests
      summary: Delete test record  by ID
      description: For valid response try integer IDs with positive integer value.         Negative
        or non-integer values will generate API errors
      operationId: deleteTest
      parameters:
        - name: recordId
          in: path
          description: ID of the well of which we need to update the record.
          required: true
          schema:
            minimum: 1.0
            type: integer
        - name: id
          in: path
          description: ID of the well test record that needs to be deleted
          required: true
          schema:
            minimum: 1.0
            type: integer
      responses:
        204:
          description: The resource was deleted successfully.
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server error.
components:
...
  schemas:
...
    AllLabs:
      type: array
      items:
        $ref: '#/components/schemas/labMeasurementResponse'
    AllLabsOfAWell:
      type: array
      items:
        $ref: '#/components/schemas/labMeasurementResponse'
...
    labMeasurementRequest:
      type: object
      properties:
        s1:
          type: integer
          format: int64
          default: 0
        s2:
          type: integer
          format: int64
          default: 0
        s3:
          type: integer
          format: int64
          default: 0
        s4:
          type: integer
          format: int64
          default: 0
        s5:
          type: integer
          format: int64
          default: 0
        remarks:
          type: string
        date:
          type: string
          format: date-time


      example:
        s1: 15.55
        s2: 9.3
        s3: 5.2
        s4: 1.3
        s5: 7
        remarks: very poor measurements
        date: 2021-03-16
    labMeasurementResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          default: 0
        s1:
          type: integer
          format: int64
          default: 0
        s2:
          type: integer
          format: int64
          default: 0
        s3:
          type: integer
          format: int64
          default: 0
        s4:
          type: integer
          format: int64
          default: 0
        s5:
          type: integer
          format: int64
          default: 0
        remarks:
          type: string
        date:
          type: string
          format: date-time
      example:
        id: 8
        s1: 15.55
        s2: 9.3
        s3: 5.2
        s4: 1.3
        s5: 7
        remarks: very poor measurements
        date: 2021-03-16=

#    allLabMeasurementResponse:
#      type: object
#      properties:
#        id:
#          type: integer
#          format: int64
#          default: 0
#        s1:
#          type: integer
#          format: int64
#          default: 0
#        s2:
#          type: integer
#          format: int64
#          default: 0
#        s3:
#          type: integer
#          format: int64
#          default: 0
#        s4:
#          type: integer
#          format: int64
#          default: 0
#        s5:
#          type: integer
#          format: int64
#          default: 0
#        remarks:
#          type: string
#        date:
#          type: string
#          format: date-time
#        wellId:
#          type: integer
#          format: int64
#      example:
#        id: 8
#        s1: 15.55
#        s2: 9.3
#        s3: 5.2
#        s4: 1.3
#        s5: 7
#        remarks: very poor measurements
#        date: 2021-03-16
#        wellId: 5
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message

    #Schema for concession Response
...
    wellTestRequest:
      type: object
      properties:
        productionDate:
          type: string
          description: "the date in which the test was measured"
          format: date-time
        tDuration:
          type: string
          description: "a value between 0-24 representing the test duration"
          default: 24
          example: 24
        gross:
          type: number
          description: "representign the gross production of the well"
          example: 1000
        net:
          type: number
          description: "representing the net production for the well"
          example: 1000
        waterCut:
          type: number
          description: "representing the watercut percentage in the production"
          example: 50
          default: 0
        gor:
          type: number
          description: "representing the Gas oil ration in th stream for the well in SCF/BBL"
          example: 1000
        gasRate:
          type: number
          description: "representing the gas rate in the production stream in SCF"
          example: 1000
          default: 100
        condensateRate:
          type: number
          description: "representing the condensate rate in the production stream"
          default: 0
          example: 100
        whp:
          type: number
          description: "representing the well head pressure for the well in PSI"
          default: 120
          example: 120
        wht:
          type: number
          description: "representing the well head temperature for the well in degree-f"
          default: 120
          example: 120
        usp:
          type: number
          description: "representing the upstream pressure for the well psi"
          default: 120
          example: 120
        ust:
          type: number
          description: "representing the upstream temperature for the well f"
          default: 120
          example: 120
        sp:
          type: number
          description: "representing the separator pressure for the well in psi"
          default: 120
          example: 120
        st:
          type: number
          description: "representing the separator temperature for the well in degree-f"
          default: 60
          example: 60
        flp:
          type: number
          description: "representing the flow line pressure for the well psi"
          default: 0
          example: 120
        flt:
          type: number
          description: "representing the flow line temperature for the well in *f "
        chockType:
          $ref: '#/components/schemas/ChockTypeEnum'
        chockSize:
          type: number
          description: "the chock size at which the the test was performed"
          default:
        h2s:
          type: number
          description: "representing the h2s ppm in the well flow"

        co2:
          type: number
          description: "representing the co2 in the well stream"
          example: 0
          default: 0
        unit:
          type: string
          description: "the test unit which performed the test"
          example: "Expro-TMU#3"
        remarks:
          type: string
          description: "Any notes reported or added during the test or recommendaton for the well"
          example: plan to perform DFL for the well
#      xml:
#        name: wellTest
    wellTestResponse:
      type: object
      properties:
        id:
          type: integer
          description: "represents the test id"
          example: 1
        productionDate:
          type: string
          description: "the date in which the test was measured"
          format: date-time
        tDuration:
          type: string
          description: "a value between 0-24 representing the test duration"
          default: 24
          example: 24
        gross:
          type: number
          description: "representign the gross production of the well"
          example: 1000
        net:
          type: number
          description: "representing the net production for the well"
          example: 1000
        waterCut:
          type: number
          description: "representing the watercut percentage in the production"
          example: 50
          default: 0
        gor:
          type: number
          description: "representing the Gas oil ration in th stream for the well in SCF/BBL"
          example: 1000
        gasRate:
          type: number
          description: "representing the gas rate in the production stream in SCF"
          example: 1000
          default: 100
        condensateRate:
          type: number
          description: "representing the condensate rate in the production stream"
          default: 0
          example: 100
        whp:
          type: number
          description: "representing the well head pressure for the well in PSI"
          default: 120
          example: 120
        wht:
          type: number
          description: "representing the well head temperature for the well in degree-f"
          default: 120
          example: 120
        usp:
          type: number
          description: "representing the upstream pressure for the well psi"
          default: 120
          example: 120
        ust:
          type: number
          description: "representing the upstream temperature for the well f"
          default: 120
          example: 120
        sp:
          type: number
          description: "representing the separator pressure for the well in psi"
          default: 120
          example: 120
        st:
          type: number
          description: "representing the separator temperature for the well in degree-f"
          default: 60
          example: 60
        flp:
          type: number
          description: "representing the flow line pressure for the well psi"
          default: 0
          example: 120
        flt:
          type: number
          description: "representing the flow line temperature for the well in *f "
        chockType:
          $ref: '#/components/schemas/ChockTypeEnum'

        chockSize:
          type: number
          description: "the chock size at which the the test was performed"
          default:
        h2s:
          type: number
          description: "representing the h2s ppm in the well flow"

        co2:
          type: number
          description: "representing the co2 in the well stream"
          example: 0
          default: 0
        unit:
          type: string
          description: "the test unit which performed the test"
          example: "Expro-TMU#3"
        remarks:
          type: string
          description: "Any notes reported or added during the test or recommendaton for the well"
#      xml:
#        name: wellTest
    AllTests:
      type: array
      items:
        $ref: '#/components/schemas/wellTestResponse'

    #budget_Actual
...
    ChockTypeEnum:
      type: string
      description: "representing the o"
      default: "fixed"
      enum:
        - fixed
        - adjustable
...